requirements.txt
transformers[torch]
pillow
flask